Fixed issue of price field and column type from double to decimal

+++ cloud-seata-jpa/order-service/src/main/java/io/github/helloworlde/order/service/impl/OrderServiceImpl.java
@@ -37,7 +37,7 @@ public class OrderServiceImpl implements OrderService {
     @Override
     public OperationResponse placeOrder(PlaceOrderRequestVO placeOrderRequestVO) {
         Integer amount = 1;
-        BigDecimal price = BigDecimal.valueOf(placeOrderRequestVO.getPrice());
+        BigDecimal price = placeOrderRequestVO.getPrice();
 
         Order order = Order.builder()
                            .userId(placeOrderRequestVO.getUserId())
